requirements.txt
pyserial==3.5
PyYAML==6.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
colorlog==6.7.0
python-dateutil==2.8.2
